sort functions consolidated
