+++ src/components/TVShows/SearchShows.js
@@ -14,14 +14,7 @@ export const SearchShows = ({ userEntries, setUserEntries, taggedShows }) => {
         () => {
             TagRepo.getTagsForUser(userId)
                 .then(result => {
-                    const sorted = result.sort((a, b) => {
-                        const tagA = a.tag.toLowerCase()
-                        const tagB = b.tag.toLowerCase()
-                        if (tagA < tagB) { return -1 }
-                        if (tagA > tagB) { return 1 }
-                        return 0 //default return value (no sorting)
-                    })
-                    setTags(sorted)
+                    setTags(sortByTag(result))
                 })
                 .then(ShowRepo.getAllStreamingServices)
                 .then(setStreamingServices)
