+++ src/components/TVShows/ShowForm.js
@@ -34,14 +34,7 @@ export const ShowForm = () => {
             //on page load, GET streaming services and tags
             TagRepo.getTagsForUser(userId)
                 .then(result => {
-                    const sorted = result.sort((a, b) => {
-                        const tagA = a.tag.toLowerCase()
-                        const tagB = b.tag.toLowerCase()
-                        if (tagA < tagB) { return -1 }
-                        if (tagA > tagB) { return 1 }
-                        return 0 //default return value (no sorting)
-                    })
-                    setTags(sorted)
+                    setTags(sortByTag(result))
                 })
                 .then(ShowRepo.getAllStreamingServices)
                 .then(setStreamingServices)
