+++ src/repositories/BookRepo.js
@@ -1,4 +1,4 @@
-import { fetchIt } from "./Fetch"
+import { fetchIt } from "./FetchAndSort"
 
 const embedTags = (book, tags) => {
     /*
