+++ src/repositories/ShowRepo.js
@@ -1,4 +1,4 @@
-import { fetchIt } from "./Fetch"
+import { fetchIt } from "./FetchAndSort"
 
 const embedTags = (show, tags) => {
     /*
