+++ src/repositories/TagRepo.js
@@ -1,4 +1,4 @@
-import { fetchIt } from "./Fetch"
+import { fetchIt } from "./FetchAndSort"
 
 export const TagRepo = {
     //GETs
